feat: refactor index.js and create util files
